+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "context-window" extension will be documented in this
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.4.3] 2025.07.11
 
-- Initial release
+- Default token parsing uses js
+- Apply js as the default token provider.
+- When multiple definitions are used, the option colors use the same method as in the vscode list, and a custom setting for the right-click selected color is added
+- Aply the same color as vscode's list in multi-definitions, add color config settings for right-click selection.
+
+## [0.4.2] 2025.07.09
+
+- Support right click to select token to view the same text
+- Support right-click to select a token for viewing identical text.